+++ NotepadKnights/CallMainScene.cs
@@ -98,7 +98,9 @@ namespace NotepadKnights
             }
         }
         public int Villiage()
-        {
+        {   
+            // 임시코드
+            int answer = 0;
             Console.WriteLine("스파르타 마을에 오신 여러분 환영합니다.");
             Console.WriteLine("이곳에서 던전으로 들어가기 전 활동을 할 수 있습니다.");
             Console.WriteLine("1. 상태 보기 / 2. 인벤토리 / 3. 상점 / 4. 전투하기 / 5. 회복하기\n");
